% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fitness.R
\name{calculate_mix_fitness}
\alias{calculate_mix_fitness}
\title{Calculate microbial fitness}
\usage{
calculate_mix_fitness(data, population_vars, strain_names, keep = NULL)
}
\arguments{
\item{data}{Data frame or data frame extension (e.g. tibble)}

\item{population_vars}{Named vector or list.}

\item{strain_names}{Named vector or list.}

\item{keep}{Optional named vector or list of column names in \code{data} to keep
in output frame (e.g. treatment variables).}
}
\value{
Object of same type as \code{data}.
}
\description{
\code{calculate_mix_fitness()} calculates several measures of microbial fitness
given a data frame with the observed abundances of two microbes before and
after the experiment.
}
