% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_smith2010}
\alias{data_smith2010}
\title{Data from smith et al. 2010}
\format{
A data frame with 28 rows and 5 columns:
\describe{
\item{exptl_block}{Experimental block. Date experiment started (YYYY-MM-DD).}
\item{initial_cells_evolved}{Initial number of GVB206.3 cells (colony-forming units per plate)}
\item{initial_cells_ancestral}{Initial number of GJV10 cells (colony-forming units per plate)}
\item{final_spores_evolved}{Final number of GVB206.3 spores (colony-forming units per plate)}
\item{final_spores_ancestral}{Final number of GJV10 spores (colony-forming units per plate)}
}
}
\source{
smith j, Van Dyken JD, and Zee PC (2010) A generalization of
Hamilton's rule for the evolution of microbial cooperation. \emph{Science}
328:1700. \url{https://doi.org/10.1126/science.1189675}
}
\usage{
data_smith2010
}
\description{
Sporulation success of \emph{Myxococcus xanthus} bacteria in mixed-genotype
fruiting bodies. GVB206.3 is an experimentally evolved genotype. GJV10 is an
ancestral genotoype.
}
\keyword{datasets}
