% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fitness.R
\name{plot_fitness}
\alias{plot_fitness}
\title{Plot strain and multilevel fitness}
\usage{
plot_fitness(data, var_names = NULL, mix_scale = "fraction")
}
\arguments{
\item{data}{Data frame of fitness values. Each row must contain data for two
microbes in the same population. Accepts data frame extensions like
\code{tibble}.}

\item{var_names}{Named character vector identifying which fitness measures
are contained in \code{data} variables. If \code{NULL}, defaults to column names
returned by \verb{calculate_fitness().} See Details.}

\item{mix_scale}{Determines x-axis scale. \code{"fraction"} uses
\code{initial_fraction_A}. \code{"ratio"} uses \code{initial_ratio_A_B} (on
\eqn{\log_{10}} scale).}
}
\description{
Draw a combined plot of strain, total group, and relative within-group
fitness against mix frequency.
}
\details{
\code{var_names} must be a named vector with the following elements (default
values shown):

\if{html}{\out{<div class="sourceCode">}}\preformatted{var_names = c(
  name_A = "name_A",
  name_B = "name_B",
  initial_fraction_A = "initial_fraction_A",
  initial_ratio_A_B = "initial_ratio_A_B",
  fitness_A = "fitness_A",
  fitness_B = "fitness_B",
  fitness_total = "fitness_total",
  fitness_ratio_A_B = "fitness_ratio_A_B"
)
}\if{html}{\out{</div>}}
}
