% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fitness.R
\name{plot_mix_fitness}
\alias{plot_mix_fitness}
\title{Plot strain and multilevel fitness}
\usage{
plot_mix_fitness(data, var_names = NULL, mix_scale = c("fraction", "ratio"))
}
\arguments{
\item{data}{Data frame of fitness values. Wide format: each row must contain
data for two microbes in the same population. Accepts data frame
extensions like \code{tibble}.}

\item{var_names}{Named character vector identifying fitness and mixing
variables in \code{data}. If \code{NULL}, defaults to column names returned by
\code{\link[=calculate_mix_fitness]{calculate_mix_fitness()}}. See Details.}

\item{mix_scale}{Determines mixing scale for x axis. When \code{"fraction"}, uses
initial frequency of strain A (proportion of total) from the
\code{initial_fraction_A} variable. When \code{"ratio"}, uses ratio of strain A
to strain B (on \eqn{\log_{10}} scale) from the \code{initial_ratio_A_B}
variable. Defaults to show both.}
}
\description{
\code{plot_mix_fitness()} shows a diagnostic overview of various fitness effects in
a dataset. Draws a combined plot of strain, total group, and relative
within-group fitness against mix frequency.
}
\details{
Expects Wrightian fitness measures like those returned by
\code{\link[=calculate_mix_fitness]{calculate_mix_fitness()}}.

\code{var_names} must be a named vector or list that includes the following
elements (shown here with default values):

\if{html}{\out{<div class="sourceCode">}}\preformatted{var_names = c(
  name_A = "name_A",
  name_B = "name_B",
  initial_fraction_A = "initial_fraction_A",
  initial_ratio_A_B = "initial_ratio_A_B",
  fitness_A = "fitness_A",
  fitness_B = "fitness_B",
  fitness_total = "fitness_total",
  fitness_ratio_A_B = "fitness_ratio_A_B"
)
}\if{html}{\out{</div>}}
}
\examples{
library(patchwork)

# Using data from smith et al (2010)
fitness_smith_2010 <- calculate_mix_fitness(
  data_smith_2010,
  var_names = c(
    initial_number_A = "initial_cells_evolved",
    initial_number_B = "initial_cells_ancestral",
    final_number_A = "final_spores_evolved",
    final_number_B = "final_spores_ancestral",
    name_A = "GVB206.3",
    name_B = "GJV10"
  )
)
plot_mix_fitness(fitness_smith_2010)

}
\seealso{
\code{\link[=calculate_mix_fitness]{calculate_mix_fitness()}}
}
