% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fitness.R
\name{plot_total_group_fitness}
\alias{plot_total_group_fitness}
\title{Plot total group fitness}
\usage{
plot_total_group_fitness(
  data,
  var_names = NULL,
  mix_scale = "fraction",
  xlab = NULL,
  ylab = NULL,
  xlim = c(NA, NA),
  ylim = c(NA, NA)
)
}
\arguments{
\item{data}{Data frame of fitness values. Accepts data frame extensions like
\code{tibble}.}

\item{var_names}{Named character vector identifying fitness and mixing
variables in \code{data}. If \code{NULL}, defaults to column names returned by
\code{\link[=calculate_mix_fitness]{calculate_mix_fitness()}}. See Details.}

\item{mix_scale}{Determines mixing scale to show on x axis. When \code{"fraction"}
(default), uses initial frequency of strain A (proportion of total) from
the \code{initial_fraction_A} variable. When \code{"ratio"}, uses ratio of strain A
to strain B (on \eqn{\log_{10}} scale) from the \code{initial_ratio_A_B}
variable.}

\item{xlab, ylab}{Optional strings giving labels for x and y axes to replace
default}

\item{xlim, ylim}{Optional x and y axis limits to replace default}
}
\value{
\code{\link[=ggplot]{ggplot()}} object that can be modified further
}
\description{
\code{plot_total_group_fitness()} draws a plot of absolute fitness for a group
containing two microbe types as a function of their initial frequency.
}
\details{
Expects Wrightian fitness measures like those returned by
\code{\link[=calculate_mix_fitness]{calculate_mix_fitness()}}. \code{var_names} must be a named vector that includes
the following elements (shown here with default values):

\if{html}{\out{<div class="sourceCode">}}\preformatted{var_names = c(
  name_A = "name_A",
  name_B = "name_B",
  initial_fraction_A = "initial_fraction_A",
  initial_ratio_A_B = "initial_ratio_A_B",
  fitness_total = "fitness_total"
)
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link[=plot_within_group_fitness]{plot_within_group_fitness()}}, \code{\link[=plot_strain_fitness]{plot_strain_fitness()}}
}
